requirements.txt
gitpython
jupyter
lizard
matplotlib
pandas
pylint
pytest
python-dateutil
python-dotenv
pydriller
rich
scipy
seaborn
tqdm